Refactor: move to domain file
